fix: add command for setup edge case (#45)
I only had Python 3.13, which satisfies the pyproject.toml >=3.12 requirement. But a library required <3.13, and then uv got confused. `uv install python` will actually install the version of Python that fits all of the requirements instead of merely looking to see if any that work already exist on your system.